requirements.txt
deepspeech==0.6.1
paho-mqtt==1.5.0
rhasspy-hermes==0.2.0
rhasspy-nlu==0.1.8
rhasspy-silence==0.1.2
